Massive changes to feedback given to users. Feedback is now given when posting, deleting posts, voting, hiding posts, adding and deleting comments, creating a new group, joining a group (and some extra error messages for when group joining has failed). Registration has now been prevented if the user had not confirmed a previous attempt, the email is just resent instead.

+++ script/grouping/confirm.php
@@ -6,29 +6,34 @@ include_once("../util/mysql.php");
 
 $group_id = $_GET["group_id"];
 
-$new_values = array("group_id"=>$group_id,"user_id"=>$user->user_id);
+if (isset($user)) {
 
-$dao = new DAO(false);
+	$new_values = array("group_id"=>$group_id,"user_id"=>$user->user_id);
 
-//Check if the user has already been added:
-$already_grouped = DataObject::select_one($dao, "grouping", array("grouping_id"), $new_values);
+	$dao = new DAO(false);
 
-if ($already_grouped == NULL) {
-	$grouping = DataObject::create($dao, "grouping", $new_values);
-	$request = DataObject::select_one($dao, "grouping_request", array("gr_id","group_id","user_id"),array("group_id"=>$group_id,"user_id"=>$user->user_id));
+	//Check if the user has already been added:
+	$already_grouped = DataObject::select_one($dao, "grouping", array("grouping_id"), $new_values);
 
-	if ($request != NULL) {
-		$request->delete();//Delete the request from the database
+	if ($already_grouped == NULL) {
+		$grouping = DataObject::create($dao, "grouping", $new_values);
+		$request = DataObject::select_one($dao, "grouping_request", array("gr_id","group_id","user_id"),array("group_id"=>$group_id,"user_id"=>$user->user_id));
 
-		if ($grouping->commit()) {
-			redirect("/group/$group_id");//Send them to the new group!
+		if ($request != NULL) {
+			$request->delete();//Delete the request from the database
+
+			if ($grouping->commit()) {
+				redirect("/",array("group_id"=>$group_id,"m"=>17));//Send them to the new group!
+			} else {
+				redirect("/?m=11");
+			}
 		} else {
-			redirect("/?m=11");
+			redirect("/?m=13"); //You have not been asked to join this group
 		}
 	} else {
-		redirect("/?m=13"); //You have not been asked to join this group
+		redirect("/",array("group_id"=>$group_id,"m"=>14)); //You are already in this group... See!
 	}
-} else {
-	redirect("/?group_id=".$group_id."&m=14"); //You are already in this group... See!
+} else {//Not logged in!
+	redirect("/welcome/?r=/script/grouping/confirm.php%3Fgroup_id%3D".$group_id);
 }
 ?>
